requirements.txt
celery[redis]
redis
requests
flask
python-dotenv
eventlet